+++ sample/conf.py
@@ -2,6 +2,11 @@
 Configuration for Sphinx.
 """
 
+import sys
+from pathlib import Path
+
+sys.path.insert(0, str(object=Path(__file__).parent))
+
 extensions = [
     "sphinxcontrib.video",
     "sphinx_iframes",
@@ -13,5 +18,6 @@ extensions = [
     "atsphinx.audioplayer",
     "sphinx_immaterial.task_lists",
     "sphinx.ext.mathjax",
+    "sphinx.ext.autodoc",
     "sphinx_notion",
 ]